accept floating point number

+++ app.py
@@ -14,7 +14,7 @@ def home():
 
 @app.route('/predict', methods=['POST'])
 def predict():
-    input_features = [int(x) for x in request.form.values()]
+    input_features = [float(x) for x in request.form.values()]
     features_value = [np.array(input_features)]
 
     features_name = ['clump_thickness', 'uniform_cell_size', 'uniform_cell_shape',
